refactor(chatbi): 移动 AskUserDialog 到 views 目录
